Updated the Feature

+++ src/events/HTOAD/Moderation/MultiLanguageFilter.js
@@ -5,6 +5,10 @@ const allowedServers = [
     '1120022058601029652' // How to Own a Dragon Server
 ];
 
+const allowedRoles = [
+    '1120033014416670895' // Bots
+];
+
 module.exports = {
     event: 'messageCreate',
     once: false,
@@ -17,12 +21,12 @@ module.exports = {
         const badWord = profanityWords.map(profanity => profanity.profanity);
 
         const containsbadWords = badWord.some(badWords => message.content.includes(badWords));
+        if (message.author.bot) return;
         if (containsbadWords || profanityFilter.check(message.content) || profanityFilter.check(message.content.toLowerCase()) || profanityFilter.check(message.content.toUpperCase()) || profanityFilter.check(message.content.charAt(0).toUpperCase() + message.content.slice(1)) || profanityFilter.check(message.content.charAt(0).toLowerCase() + message.content.slice(1))) {
             if (allowedServers.includes(message.guild.id)) {
                 const cleanMessage = profanityFilter.clean(message.content);
                 await message.delete();
                 await message.channel.send(`${message.author} said: ${cleanMessage}`);
-                await message.author.send(`Please do not use profanity in the server!`);
                 const logChannelId = '1131214666757058654'; // How to Own a Dragon message-automod channel ID
                 const logChannel = await client.channels.fetch(logChannelId);
